Raw move of translation code into translator
